In "core" project moved "Vertx" starting classes and "Validation" trait from "utils" package to "services" one
