+++ services/authentication/src/main/scala/it/cwmp/services/authentication/AuthenticationServiceVerticle.scala
@@ -3,9 +3,10 @@ package it.cwmp.services.authentication
 import io.netty.handler.codec.http.HttpResponseStatus._
 import io.vertx.core.Handler
 import io.vertx.scala.ext.web.{Router, RoutingContext}
+import it.cwmp.services.VertxServer
 import it.cwmp.services.authentication.ServerParameters._
 import it.cwmp.utils.Utils.{httpStatusNameToCode, stringToOption}
-import it.cwmp.utils.{HttpUtils, Logging, VertxServer}
+import it.cwmp.utils.{HttpUtils, Logging}
 
 import scala.concurrent.Future
 import scala.util.{Failure, Success}
