+++ services/room-receiver/src/test/scala/it/cwmp/services/roomreceiver/RoomReceiverServiceVerticleTest.scala
@@ -5,11 +5,11 @@ import io.vertx.lang.scala.json.Json
 import io.vertx.scala.ext.web.client.WebClientOptions
 import it.cwmp.exceptions.HTTPException
 import it.cwmp.model.Participant.Converters._
+import it.cwmp.services.VertxClient
 import it.cwmp.services.roomreceiver.ServerParameters._
 import it.cwmp.services.testing.roomreceiver.RoomReceiverWebTesting
 import it.cwmp.testing.{FutureMatchers, HttpMatchers}
 import it.cwmp.utils.Utils.httpStatusNameToCode
-import it.cwmp.utils.VertxClient
 
 import scala.util.{Failure, Success}
