+++ services/rooms/src/main/scala/it/cwmp/services/rooms/RoomsServiceMain.scala
@@ -1,7 +1,8 @@
 package it.cwmp.services.rooms
 
+import it.cwmp.services.VertxInstance
 import it.cwmp.services.wrapper.{AuthenticationApiWrapper, RoomReceiverApiWrapper}
-import it.cwmp.utils.{Logging, VertxInstance}
+import it.cwmp.utils.Logging
 
 import scala.util.{Failure, Success}
